mean: Only deploy one HaProxy container
Before, we deployed a HaProxy container on each machine so that users
could use any machine IP to access the deployment. However, we are going
to remove support for placement rules soon, and it's not difficult for
users to get the public IP of HaProxy container from the `quilt show`
output.